# docs/swagger.yaml
definitions:
  dtos.AIDayRequest:
    properties:
      dayNumber:
        type: integer
      workout:
        $ref: '#/definitions/dtos.AIWorkoutRequest'
    required:
    - dayNumber
    - workout
    type: object
  dtos.AIExerciseRequest:
    properties:
      exerciseId:
        type: integer
      sets:
        items:
          $ref: '#/definitions/dtos.AISetRequest'
        type: array
    required:
    - exerciseId
    - sets
    type: object
  dtos.AISetRequest:
    properties:
      reps:
        type: integer
      rest:
        type: integer
    required:
    - reps
    - rest
    type: object
  dtos.AIWorkoutRequest:
    properties:
      exercises:
        items:
          $ref: '#/definitions/dtos.AIExerciseRequest'
        type: array
      name:
        type: string
    required:
    - exercises
    - name
    type: object
  dtos.AddTrainingDayRequest:
    properties:
      dayNumber:
        minimum: 1
        type: integer
      notes:
        type: string
      workoutId:
        type: integer
    required:
    - dayNumber
    - workoutId
    type: object
  dtos.AssignTrainingPlanRequest:
    properties:
      startDate:
        type: string
      userId:
        type: integer
    required:
    - startDate
    - userId
    type: object
  dtos.CloneTrainingPlanRequest:
    properties:
      targetUserId:
        type: integer
    required:
    - targetUserId
    type: object
  dtos.CreateTrainingPlanFromAIRequest:
    properties:
      days:
        items:
          $ref: '#/definitions/dtos.AIDayRequest'
        type: array
      durationDays:
        type: integer
      name:
        type: string
      userId:
        type: integer
    required:
    - days
    - durationDays
    - name
    - userId
    type: object
  dtos.CreateTrainingPlanRequest:
    properties:
      description:
        type: string
      durationDays:
        minimum: 1
        type: integer
      isTemplate:
        type: boolean
      name:
        type: string
    required:
    - durationDays
    - name
    type: object
  dtos.UpdateDayCompletionRequest:
    properties:
      isCompleted:
        type: boolean
    type: object
  models.NutritionPlan:
    properties:
      createdAt:
        type: string
      dailyCalories:
        type: number
      dailyCarbs:
        type: number
      dailyFats:
        type: number
      dailyProtein:
        type: number
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      objective:
        description: e.g., weight_loss, muscle_gain, maintenance
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  usecases.GenerateNutritionPlanInput:
    properties:
      activityLevel:
        description: 1.2 (sedentary) to 1.9 (heavy)
        type: number
      age:
        type: integer
      gender:
        description: male, female
        type: string
      height:
        description: cm
        type: number
      objective:
        description: weight_loss, muscle_gain, maintenance
        type: string
      userId:
        type: integer
      weight:
        description: kg
        type: number
    type: object
info:
  contact: {}
paths:
  /internal/training-plans/ai:
    post:
      consumes:
      - application/json
      description: Stores a plan generated by the AI service. Validates exercises
        and assigns to user.
      parameters:
      - description: AI Plan data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateTrainingPlanFromAIRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create a training plan from AI (INTERNAL ONLY)
      tags:
      - Internal
  /private/nutrition-plans/generate:
    post:
      consumes:
      - application/json
      description: Calculates macros and creates a plan based on user metrics and
        goals.
      parameters:
      - description: Plan generation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecases.GenerateNutritionPlanInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NutritionPlan'
      summary: Generate nutritional plan
      tags:
      - nutrition
  /private/training-plans:
    post:
      consumes:
      - application/json
      description: Creates a new training plan. Accessible by trainers (role 3) and
        regular users (role 4).
      parameters:
      - description: Plan data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateTrainingPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a training plan
      tags:
      - TrainingPlans
  /private/training-plans/:id/clone:
    post:
      consumes:
      - application/json
      description: Clones an existing plan (template or other) and assigns it to a
        specific user.
      parameters:
      - description: Plan ID to clone
        in: path
        name: id
        required: true
        type: integer
      - description: Clone data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CloneTrainingPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Clone a training plan for a user
      tags:
      - TrainingPlans
  /private/training-plans/:id/days/:dayId/complete:
    patch:
      consumes:
      - application/json
      description: Marks a training day as completed or not.
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Day ID
        in: path
        name: dayId
        required: true
        type: integer
      - description: Progress data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateDayCompletionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update training day progress
      tags:
      - TrainingPlans
  /private/training-plans/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a training plan with its days and workouts. Users can
        only see their assigned plans.
      parameters:
      - description: Training Plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get training plan by ID
      tags:
      - TrainingPlans
  /private/training-plans/{id}/assign:
    post:
      consumes:
      - application/json
      description: Assigns an existing training plan to a user. Only accessible by
        trainers (role 3).
      parameters:
      - description: Training Plan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Assignment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.AssignTrainingPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign training plan to a user (TRAINER ONLY)
      tags:
      - TrainingPlans
  /private/training-plans/{id}/days:
    post:
      consumes:
      - application/json
      description: Adds a workout day to an existing training plan. DayNumber must
        be within the plan's duration.
      parameters:
      - description: Training Plan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Day data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.AddTrainingDayRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add a day to a training plan
      tags:
      - TrainingPlans
  /private/training-plans/adapt:
    post:
      consumes:
      - application/json
      description: Analyzes user progress and creates an adapted version of the latest
        plan if completion is high.
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Adapt training plan based on progress
      tags:
      - TrainingPlans
  /private/training-plans/user/{userId}:
    get:
      consumes:
      - application/json
      description: Returns all training plans assigned to a specific user. Regular
        users can only query their own.
      parameters:
      - description: Target User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get training plans for a user
      tags:
      - TrainingPlans
  /public/exercises:
    get:
      consumes:
      - application/json
      description: Retrieve all exercises. Optionally filter by bodyPart and target.
      parameters:
      - description: Filter by body part
        in: query
        name: bodyPart
        type: string
      - description: Filter by target
        in: query
        name: target
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get list of exercises
      tags:
      - Exercises
  /public/exercises/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific exercise using its unique ID.
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get exercise by ID
      tags:
      - Exercises
  /public/exercises/import:
    post:
      consumes:
      - application/json
      description: Fetch exercises from external API and store them locally. Highly
        idempotent.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Manually import exercises
      tags:
      - Exercises
  /public/foods:
    get:
      consumes:
      - application/json
      description: Retrieve a list of foods filtered by name.
      parameters:
      - description: Food name to search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Search foods
      tags:
      - Nutrition
  /public/foods/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific food item.
      parameters:
      - description: Food ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get food by ID
      tags:
      - Nutrition
  /public/foods/import:
    post:
      consumes:
      - application/json
      description: Triggers a one-time import of basic food categories from USDA API.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Manual import of foods from USDA
      tags:
      - Nutrition
  /public/workouts/{id}/full:
    get:
      consumes:
      - application/json
      description: Retrieve a workout with its exercises and sets in a frontend-friendly
        structure.
      parameters:
      - description: Workout ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get full workout details
      tags:
      - Workouts
swagger: "2.0"